requirements.txt
streamlit==1.52.1
pandas>=2.1
numpy>=1.26
refinitiv-data==1.6.2
httpx<0.27.0
